fix bug of being able to add items in cart with quantity more than the quantity in stock
