+++ public/css/inventory.css
@@ -282,8 +282,9 @@ input[name="cartItem"] {
 
 #checkout {
   flex: 0 0 500px;
-  padding: 10px;
+  padding: 20px;
   background-color: #f8f8f8;
+  border: 2px solid #4CAF50;
   border-radius: 5px;
   box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
 }
