+++ views/private_views/dashboard.ejs
@@ -1,6 +1,7 @@
 <%- include('private_partials/privateHeader') %>
         <main class="mdl-layout__content">
            <h1>Dashboard</h1>
+
         </main>
     </div>
 </body>
