+++ views/public_views/content.ejs
@@ -21,7 +21,7 @@
                         <div class="content-card">
                             <div class="content-head">
                                 <h3><%= content.title %></h3>
-                                <p class="date" data-created="<%= content.created_at %>"></p>
+                                <p class="date" data-created="<%= content.updated_at %>"></p>
                             </div>
                             <div class="content-body"><%- content.content.html %></div>
                         </div>
